added WorkManagerUtils, creation of job & enque to PreplannedMapAreasState
