+++ toolkit/offline/src/main/java/com/arcgismaps/toolkit/offline/OfflineMapAreas.kt
@@ -32,6 +32,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.toMutableStateList
@@ -40,8 +41,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import androidx.work.WorkManager
 import com.arcgismaps.toolkit.offline.preplanned.PreplannedMapAreas
-import com.arcgismaps.toolkit.offline.preplanned.Status
+import kotlinx.coroutines.launch
 
 /**
  * Take a web map offline by downloading map areas.
@@ -58,7 +60,13 @@ public fun OfflineMapAreas(
     val initializationStatus by offlineMapState.initializationStatus
 
     LaunchedEffect(offlineMapState) {
-        offlineMapState.initialize(scope, context)
+        offlineMapState.apply {
+            // TODO: Use singleton/centralized manager
+            workManager = WorkManager.getInstance(context)
+            getExternalFilesDirPath = context.getExternalFilesDir(null)?.path.toString()
+
+        }
+        offlineMapState.initialize()
     }
 
     Surface(
@@ -84,13 +92,14 @@ public fun OfflineMapAreas(
 
             else -> {
                 if (offlineMapState.mode == OfflineMapMode.Preplanned) {
+                    val preplannedAreas = remember { offlineMapState.preplannedMapAreaStates.toMutableStateList() }
                     PreplannedMapAreas(
-                        preplannedMapAreaStates = offlineMapState.preplannedMapAreaStates.toMutableStateList(),
+                        preplannedMapAreaStates = preplannedAreas,
                         modifier = modifier,
-                        onDownloadButtonClicked = { selectedState ->
-                            if (selectedState.status.allowsDownload) {
-                                offlineMapState.preplannedMapAreaStates.find { it == selectedState }?.apply {
-                                    offlineMapState.takePreplannedMapOffline(this)
+                        onDownloadButtonClicked = { preplannedMapAreaState ->
+                            if (preplannedMapAreaState.status.allowsDownload) {
+                                scope.launch {
+                                    preplannedMapAreaState.takePreplannedMapOffline()
                                 }
                             }
                         }
